pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
]
